fix(bridge): converge Shape.isValidForBoolean* onto isBooleanValid* (#1297)
Four public Shape methods (isValidForBoolean, isValidForBoolean(with:),
isBooleanValid, isBooleanValidWith) reached the same BRepAlgoAPI_Check
op through two independent bridge implementations. The older pair's
bridge function (OCCTShapeBooleanCheck, OCCTBridge_Modeling.mm) guarded
its OCCTShapeRef argument before dereferencing; the newer pair
(OCCTShapeBooleanCheckSingle/Pair, OCCTBridge_Healing.mm) did not.

- OCCTShapeBooleanCheckSingle/Pair now guard all OCCTShapeRef parameters
  (if (!shape) return false;), matching the older sibling's precedent.
- OCCTShapeBooleanCheck is deleted; isValidForBoolean/isValidForBoolean(with:)
  now forward onto isBooleanValid()/isBooleanValidWith(_:) at their
  existing implicit defaults (testSmallEdges: true, testSelfInterference:
  true, operation: 0/BOPAlgo_UNKNOWN, all confirmed to match
  BRepAlgoAPI_Check's own C++ constructor defaults), so behavior is
  unchanged.
- OCCTBridge.h's class index updated; docs/reference cross-linked.
- BRepAlgoCheckTests.swift merged into BooleanCheckTests.swift, plus two
  new parity tests pinning that the two spellings agree at matching
  defaults.
