--- Tests/OCCTModelingTests/BRepAlgoCheckTests.swift
@@ -1,46 +1,0 @@
-import Testing
-import simd
-
-@testable import OCCTSwift
-
-@Suite("BRepAlgoAPI_Check")
-struct BRepAlgoCheckTests {
-    @Test func singleShapeValid() {
-        let box = Shape.box(width: 10, height: 20, depth: 30)
-        if let b = box {
-            #expect(b.isBooleanValid())
-        }
-    }
-
-    @Test func sphereValid() {
-        let sphere = Shape.sphere(radius: 5)
-        if let s = sphere {
-            #expect(s.isBooleanValid())
-        }
-    }
-
-    @Test func pairShapesValidForFuse() {
-        let box = Shape.box(width: 10, height: 20, depth: 30)
-        let sphere = Shape.sphere(radius: 5)
-        if let b = box, let s = sphere {
-            // operation 2 = BOPAlgo_FUSE
-            #expect(b.isBooleanValidWith(s, operation: 2))
-        }
-    }
-
-    @Test func pairShapesValidForCut() {
-        let box = Shape.box(width: 10, height: 20, depth: 30)
-        let sphere = Shape.sphere(radius: 5)
-        if let b = box, let s = sphere {
-            // operation 3 = BOPAlgo_CUT
-            #expect(b.isBooleanValidWith(s, operation: 3))
-        }
-    }
-
-    @Test func singleShapeNoSelfInterference() {
-        let box = Shape.box(width: 10, height: 20, depth: 30)
-        if let b = box {
-            #expect(b.isBooleanValid(testSmallEdges: false, testSelfInterference: true))
-        }
-    }
-}
