+++ docs/reference/Shape-Healing.md
@@ -1710,7 +1710,8 @@ public var isValidForBoolean: Bool { get }
 ```
 
 - **Returns:** `true` if the shape passes the OCCT boolean readiness check.
-- **OCCT:** `BRepAlgoAPI_Check` (via `OCCTShapeBooleanCheck`).
+- **OCCT:** `BRepAlgoAPI_Check` (via `OCCTShapeBooleanCheckSingle`; a convenience spelling of
+  `isBooleanValid()` at its defaults, see [Document-Transforms.md](Document-Transforms.md#shapeisbooleanvalidtestsmalledgestestselfinterference)).
 - **Example:**
   ```swift
   guard shape.isValidForBoolean else { return }
@@ -1729,7 +1730,8 @@ public func isValidForBoolean(with other: Shape) -> Bool
 
 - **Parameters:** `other`, the second shape to check compatibility with.
 - **Returns:** `true` if both shapes are suitable for boolean operations together.
-- **OCCT:** `BRepAlgoAPI_Check` (via `OCCTShapeBooleanCheck`).
+- **OCCT:** `BRepAlgoAPI_Check` (via `OCCTShapeBooleanCheckPair`; a convenience spelling of
+  `isBooleanValidWith(_:operation:)` at its defaults, see [Document-Transforms.md](Document-Transforms.md#shapeisbooleanvalidwith_operationtestsmalledgestestselfinterference)).
 - **Example:**
   ```swift
   guard partA.isValidForBoolean(with: partB) else { return }
